Refactoring to CMake project

+++ include/c_parser.h
@@ -8,5 +8,5 @@ template <typename T>
 using function_t = std::function<bool(parse_tree<T>&)>;
 
 extern const std::string c_parser_strings[];
-extern const size_t c_parser_size;
+extern const uint64_t c_parser_size;
 extern const unsigned char c_parser_table[];
